Refactor tweet-service and notification-service.

+++ commons-new/src/main/java/com/gmail/merikbest2015/dto/TweetResponse.java
@@ -29,10 +29,10 @@ public class TweetResponse {
     private List<ImageResponse> images;
     private QuoteTweetResponse quoteTweet;
     private PollResponse poll;
-    private Integer retweetsCount;
-    private Integer likedTweetsCount;
-    private Integer repliesCount;
-    private Integer quotesCount;
+    private Long retweetsCount;
+    private Long likedTweetsCount;
+    private Long repliesCount;
+    private Long quotesCount;
 
     @JsonProperty("isDeleted")
     private boolean isDeleted;
